+++ CHANGES.md
@@ -1,5 +1,10 @@
 # storehaus #
 
+### Version 0.0.3 ###
+
+* Change up `Store` interface's multiGet.
+* ChannelBuffer bijection.
+
 ### Version 0.0.2 ###
 
 * `storehaus-algebra`